backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
supabase==0.7.1
pandas==1.5.3
pillow==9.4.0
tensorflow==2.12.0
numpy==1.23.5
python-dotenv==1.0.0